+++ ui-src/src/flex-hooks/notifications/Callback.tsx
@@ -3,20 +3,16 @@ import { StringTemplates } from '../strings/Callback';
 
 // Export the notification IDs an enum for better maintainability when accessing them elsewhere
 export enum CallbackNotification {
+  ErrorCallingCustomer = 'CallbackErrorCallingCustomer',
+  ErrorRequeuingCallbackTask = 'CallbackErrorRequeuingCallbackTask',
   OutboundDialingNotEnabled = 'CallbackOutboundDialingNotEnabled',
-  ErrorCallBackAndVoicemail = 'ErrorCallBackAndVoicemail',
 }
-
-export default (flex: typeof Flex, manager: Flex.Manager) => {
-  errorNotification(flex, manager);
-  outboundDialingNotEnabled(flex, manager);
-};
 
 function errorNotification(flex: typeof Flex, manager: Flex.Manager) {
   flex.Notifications.registerNotification({
-    id: CallbackNotification.ErrorCallBackAndVoicemail,
+    id: CallbackNotification.ErrorCallingCustomer,
     type: Flex.NotificationType.error,
-    content: StringTemplates.ErrorCallBackAndVoicemail,
+    content: StringTemplates.ErrorCallingCustomer,
   });
 }
 
@@ -27,3 +23,8 @@ function outboundDialingNotEnabled(flex: typeof Flex, manager: Flex.Manager) {
     content: StringTemplates.OutboundDialingNotEnabled,
   });
 }
+
+export default (flex: typeof Flex, manager: Flex.Manager) => {
+  errorNotification(flex, manager);
+  outboundDialingNotEnabled(flex, manager);
+};