+++ ui-src/src/flex-hooks/paste-elements/index.ts
@@ -12,6 +12,12 @@ const customPasteElements = {
   C_AND_V_CONTENT_HEADING: {
     marginBottom: 'space0',
   },
+  C_AND_V_ICON: {
+    width: '100%',
+    height: '100%',
+    minWidth: 'sizeSquare90',
+    padding: 'space20',
+  },
 } as { [key: string]: PasteCustomCSS };
 
 export default customPasteElements;