+++ ui-src/src/types/ServiceConfiguration.ts
@@ -1,9 +1,7 @@
 import * as Flex from '@twilio/flex-ui';
 
 export default interface CallbackAndVoicemailConfig {
-  enabled: boolean;
   allow_requeue: boolean;
-  max_attempts: number;
   auto_select_task: boolean;
 }